+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Changed
 
+- **Use `pydantic` instead of `confz`.**
 - The pre-commit hook will not run with `--quiet`.
 - Updated the settings system in `__main__.py`.